feat: Added tsvReader to help seed db with info from vegetable_conversion.tsv
seed/tsvReader.go has been added to help seed the database with vegetable information.
The vegetable_conversion.tsv file contains information about vegetables and suggested
portions for each. The tsvReader is now used by the seeder to populate the db with it.
The table being populated with that data is the food_items table.

Tests were also updated to account for the amount of data that will now be seeded
prior to running tests.